% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{add}
\alias{add}
\title{Add two numbers}
\usage{
add(x, y)
}
\arguments{
\item{x}{A numeric value}

\item{y}{A numeric value}
}
\value{
The sum of x and y
}
\description{
Add two numbers
}
\examples{
add(2, 3)
}
