% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geodata}
\alias{geodata}
\alias{EG_2020_06_24}
\alias{EG_Bonn}
\alias{EG_DMW}
\alias{EG_Muenster}
\alias{EG_Paderborn}
\alias{EG_Siegen}
\alias{Wenkerorte}
\title{Geodaten von Erhebungsgebieten}
\format{
Ein sf-Objekt (simple features) mit Multipolygon oder Point Geometrien.
Die genaue Struktur variiert je nach Datensatz.

An object of class \code{sf} (inherits from \code{data.frame}) with 1 rows and 31 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 1 rows and 9 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 4 rows and 43 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 1 rows and 7 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 1 rows and 9 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 1 rows and 31 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 3815 rows and 16 columns.
}
\source{
Ursprünglich als GeoJSON-Dateien im Ordner \code{data} verfügbar
}
\usage{
EG_2020_06_24

EG_Bonn

EG_DMW

EG_Muenster

EG_Paderborn

EG_Siegen

Wenkerorte
}
\description{
Diese Datensätze enthalten Geodaten (sf-Objekte) mit Erhebungsgebieten
(EG) und Wenker-Orte als GeoJSON-Daten.
}
\details{
\itemize{
\item \code{EG_2020_06_24}: Erhebungsgebiete vom 24. Juni 2020
\item \code{EG_Bonn}: Erhebungsgebiete im Raum Bonn
\item \code{EG_DMW}: Erhebungsgebiete des DMW (Deutsch Mittelwort)
\item \code{EG_Muenster}: Erhebungsgebiete im Raum Münster
\item \code{EG_Paderborn}: Erhebungsgebiete im Raum Paderborn
\item \code{EG_Siegen}: Erhebungsgebiete im Raum Siegen
\item \code{Wenkerorte}: Kombinierte Wenker-Orte aus allen Regionen (Bonn, Münster, Paderborn, Siegen)
}
}
\examples{
\dontrun{
  library(sf)

  # Erhebungsgebiet Bonn
  plot(EG_Bonn)

  # Wenker-Orte
  plot(Wenkerorte)
}

}
\keyword{datasets}
