% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{voronoi}
\alias{voronoi}
\alias{WenkerOrte_Bonn_Voronoi}
\alias{WenkerOrte_Muenster_Voronoi}
\alias{WenkerOrte_Paderborn_Voronoi}
\alias{WenkerOrte_Siegen_Voronoi}
\alias{Wenkerorte_Voronoi}
\title{Voronoi-Zerlegungen der Wenker-Orte}
\format{
Ein sf-Objekt (simple features) mit Polygon-Geometrien.

An object of class \code{sf} (inherits from \code{data.frame}) with 1309 rows and 16 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 546 rows and 16 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 776 rows and 16 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 1183 rows and 16 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 3814 rows and 16 columns.
}
\source{
Erzeugt aus den Wenker-Orte-Punktdaten mittels \code{sf::st_voronoi()},
siehe \code{data-raw/create_voronoi.R}.
}
\usage{
WenkerOrte_Bonn_Voronoi

WenkerOrte_Muenster_Voronoi

WenkerOrte_Paderborn_Voronoi

WenkerOrte_Siegen_Voronoi

Wenkerorte_Voronoi
}
\description{
Diese Datensätze enthalten Voronoi-Polygone (sf-Objekte), die aus den
Wenker-Orte-Punkten der jeweiligen Region erzeugt wurden. Jedes Polygon
trägt die Attribute des nächstgelegenen Wenker-Ortes.
}
\details{
\itemize{
\item \code{WenkerOrte_Bonn_Voronoi}: Voronoi-Zerlegung der Wenker-Orte im Raum Bonn
\item \code{WenkerOrte_Muenster_Voronoi}: Voronoi-Zerlegung der Wenker-Orte im Raum Münster
\item \code{WenkerOrte_Paderborn_Voronoi}: Voronoi-Zerlegung der Wenker-Orte im Raum Paderborn
\item \code{WenkerOrte_Siegen_Voronoi}: Voronoi-Zerlegung der Wenker-Orte im Raum Siegen
\item \code{Wenkerorte_Voronoi}: Kombinierte Voronoi-Zerlegung aller Wenker-Orte
}
}
\examples{
\dontrun{
  library(sf)
  plot(WenkerOrte_Bonn_Voronoi)
}

}
\keyword{datasets}
